backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
azure-storage-blob==12.22.0
Pillow==10.4.0
PyMuPDF==1.24.7
requests==2.32.3
aiohttp==3.10.3
boto3==1.34.162
botocore==1.34.162

